+++ Assets/Scripts/Weapon/Bullet.cs
@@ -34,7 +34,7 @@ using UnityEngine;
         if (collision.gameObject.CompareTag("Monster"))
         {
             Monster monster = collision.gameObject.GetComponent<Monster>();
-            monster.Damaged(damage);
+            monster.Damaged((int)damage);
         }
     }
